(feature) Initial commit of nochlehead
Nochlehead is a storage service for compiled pinochle games. It accepts
compiled games as json input and uses mongoDB for now as its backend to
store the games. Nochlehead will also be a means for validating play,
only accepting games for which valid moves were played, or marking them
as legitimate cheats when storing them. Eventually it is intended to
also be a service to interface with the database to get games of a
certain property from the database for analytics.